requirements.txt
Flask
Flask-UUID
Flask-SQLAlchemy
psycopg2
pandas
numpy
tqdm
biopython
natsort
